requirements.txt
fastapi
uvicorn[standard]
sqlmodel
psycopg2-binary
python-jose[cryptography]
bcrypt
python-dotenv
passlib[bcrypt]